requirements.txt
pillow==11.2.1
pypdf==5.5.0
pytesseract==0.3.13